api/requirements.txt
fastapi
uvicorn
pydantic
databases
aiomysql
sqlalchemy
passlib
python-jose
cryptography
python-multipart